+++ CHANGELOG.md
@@ -5,7 +5,25 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.6.0 - Unreleased
+## Version 0.6.1 - Unreleased
+
+### Added:
+* Add `remember_parent` option to CocoDataset.union
+* Add `CocoDataset.coerce_multiple` and `load_multiple` loader with parallel processing.
+* Add parallel IO to `kwcoco union`
+* Add util parallel and eval
+
+### Fixed:
+* Issue in detection metrics when classes was not specified. Now uses the true
+  detections to infer them if possible.
+* Detection metric assign workers was never used
+* Subsets of metrics now work when TN is not available.
+
+### Changed
+* CLI backend changed to use scriptconfig ModalCLI
+
+
+## Version 0.6.0 - Released 2023-04-04
 
 ### Changed
 * CocoImage and CocoAsset use the new AliasDictProxy interface to smoothly
@@ -41,7 +59,7 @@ Bug in `CocoImage.add_asset` where json serialization was broken
 * Deprecate `CocoImage.add_auxiliary_item` in favor of `CocoImage.add_asset`
 
 ### Added
-* Added `annots` property to CocoImge
+* Added `annots` property to CocoImage
 * Added `add_annotation` to `CocoImage`
 * Added `autofix` option `kwcoco reroot`